se-book-adapted: open in a new tab with a different nav bar
Website has a flat version and a nested verstion but the book has only one version.
As it takes a lot of work to create two version for the book, let's open the book
in a new tab with a different nav bar that does not mention nested or flat.